# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Seller {
  id: number;
  name: string;
  email: string;
  phone: string;
  walletAddress: string;
  totalEarnings?: number;
  createdAt: string;
}

export interface SellerInput {
  name: string;
  email: string;
  phone: string;
  password: string;
  walletAddress: string;
}

export interface SellerLogin {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  seller: Seller;
}

export type DealStatus = typeof DealStatus[keyof typeof DealStatus];


export const DealStatus = {
  created: 'created',
  locked: 'locked',
  dispatched: 'dispatched',
  delivered: 'delivered',
  settled: 'settled',
  disputed: 'disputed',
} as const;

export type DisputeReason = typeof DisputeReason[keyof typeof DisputeReason];


export const DisputeReason = {
  item_never_arrived: 'item_never_arrived',
  wrong_damaged_item: 'wrong_damaged_item',
  incomplete_service: 'incomplete_service',
} as const;

export type DisputeStatus = typeof DisputeStatus[keyof typeof DisputeStatus];


export const DisputeStatus = {
  open: 'open',
  counter_submitted: 'counter_submitted',
  resolved_refund: 'resolved_refund',
  resolved_seller: 'resolved_seller',
  escalated: 'escalated',
} as const;

export interface Dispute {
  id: number;
  dealId: number;
  reason: DisputeReason;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  /** @nullable */
  counterProofUrl?: string | null;
  /** @nullable */
  counterProofDescription?: string | null;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface Deal {
  id: number;
  code: string;
  shortUrl: string;
  itemName: string;
  /** @nullable */
  description: string | null;
  price: number;
  feeAmount?: number;
  sellerPayout?: number;
  deliveryWindowHours: number;
  status: DealStatus;
  sellerId: number;
  /** @nullable */
  buyerPhone?: string | null;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  dispatchedAt?: string | null;
  /** @nullable */
  deliveryDeadline?: string | null;
  /** @nullable */
  settledAt?: string | null;
  dispute?: Dispute;
  createdAt: string;
}

export interface DashboardStats {
  totalDeals: number;
  settledDeals: number;
  activeDeals: number;
  disputedDeals: number;
  totalEarnings: number;
  pendingEarnings: number;
  recentSettlements: Deal[];
}

export type DealPublicStatus = typeof DealPublicStatus[keyof typeof DealPublicStatus];


export const DealPublicStatus = {
  created: 'created',
  locked: 'locked',
  dispatched: 'dispatched',
  delivered: 'delivered',
  settled: 'settled',
  disputed: 'disputed',
} as const;

export interface DealPublic {
  id: number;
  code: string;
  itemName: string;
  /** @nullable */
  description?: string | null;
  price: number;
  deliveryWindowHours: number;
  status: DealPublicStatus;
  sellerName: string;
  createdAt: string;
}

export interface DealList {
  deals: Deal[];
  total: number;
}

export interface DealInput {
  itemName: string;
  description?: string;
  price: number;
  deliveryWindowHours?: number;
}

export type PaymentInputPaymentMethod = typeof PaymentInputPaymentMethod[keyof typeof PaymentInputPaymentMethod];


export const PaymentInputPaymentMethod = {
  momo: 'momo',
  card: 'card',
} as const;

export interface PaymentInput {
  buyerName: string;
  buyerPhone: string;
  paymentMethod?: PaymentInputPaymentMethod;
}

export type DisputeInputReason = typeof DisputeInputReason[keyof typeof DisputeInputReason];


export const DisputeInputReason = {
  item_never_arrived: 'item_never_arrived',
  wrong_damaged_item: 'wrong_damaged_item',
  incomplete_service: 'incomplete_service',
} as const;

export interface DisputeInput {
  reason: DisputeInputReason;
  description?: string;
  evidenceUrl?: string;
}

export interface CounterProofInput {
  counterProofUrl: string;
  counterProofDescription?: string;
}

export interface DisputeAdmin {
  id: number;
  dealId: number;
  deal: Deal;
  reason: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  /** @nullable */
  counterProofUrl?: string | null;
  /** @nullable */
  counterProofDescription?: string | null;
  status: string;
  createdAt: string;
}

export type DisputeResolutionDecision = typeof DisputeResolutionDecision[keyof typeof DisputeResolutionDecision];


export const DisputeResolutionDecision = {
  favor_buyer: 'favor_buyer',
  favor_seller: 'favor_seller',
} as const;

export interface DisputeResolution {
  decision: DisputeResolutionDecision;
  resolution?: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  deal_created: 'deal_created',
  payment_locked: 'payment_locked',
  dispatched: 'dispatched',
  delivered: 'delivered',
  settled: 'settled',
  dispute_raised: 'dispute_raised',
  dispute_resolved: 'dispute_resolved',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  dealId: number;
  itemName: string;
  amount: number;
  /** @nullable */
  buyerName?: string | null;
  timestamp: string;
}

export type ListDealsParams = {
status?: ListDealsStatus;
limit?: number;
offset?: number;
};

export type ListDealsStatus = typeof ListDealsStatus[keyof typeof ListDealsStatus];


export const ListDealsStatus = {
  created: 'created',
  locked: 'locked',
  dispatched: 'dispatched',
  delivered: 'delivered',
  settled: 'settled',
  disputed: 'disputed',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};