# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SETTLE escrow platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List all deals for the authenticated seller
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [created, locked, dispatched, delivered, settled, disputed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealList"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a new deal/escrow link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Deal created with link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete an unpaid deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /deals/link/{code}:
    get:
      operationId: getDealByCode
      tags: [deals]
      summary: Get deal by short code (buyer portal)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deal for buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealPublic"
        "404":
          description: Not found
  /deals/{id}/pay:
    post:
      operationId: initiatePayment
      tags: [deals]
      summary: Buyer initiates payment (locks funds)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/dispatch:
    post:
      operationId: markDispatched
      tags: [deals]
      summary: Seller marks item as dispatched
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/confirm:
    post:
      operationId: confirmDelivery
      tags: [deals]
      summary: Buyer confirms delivery and releases funds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Funds released, deal settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
...
  /admin/disputes/{id}/resolve:
    post:
      operationId: resolveDispute
      tags: [admin]
      summary: Admin resolves a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeResolution"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
components:
  schemas:
...
    Deal:
      type: object
      required: [id, code, shortUrl, itemName, description, price, deliveryWindowHours, status, sellerId, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        shortUrl:
          type: string
        itemName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        feeAmount:
          type: number
        sellerPayout:
          type: number
        deliveryWindowHours:
          type: integer
        status:
          type: string
          enum: [created, locked, dispatched, delivered, settled, disputed]
        sellerId:
          type: integer
        buyerPhone:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
        dispatchedAt:
          type: ["string", "null"]
        deliveryDeadline:
          type: ["string", "null"]
        settledAt:
          type: ["string", "null"]
        dispute:
          $ref: "#/components/schemas/Dispute"
        createdAt:
          type: string
    DealPublic:
      type: object
      required: [id, code, itemName, price, deliveryWindowHours, status, sellerName, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        itemName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        deliveryWindowHours:
          type: integer
        status:
          type: string
          enum: [created, locked, dispatched, delivered, settled, disputed]
        sellerName:
          type: string
        createdAt:
          type: string
    DealList:
      type: object
      required: [deals, total]
      properties:
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
    DealInput:
      type: object
      required: [itemName, price]
      properties:
        itemName:
          type: string
        description:
          type: string
        price:
          type: number
        deliveryWindowHours:
          type: integer
          default: 48
    PaymentInput:
      type: object
      required: [buyerName, buyerPhone]
      properties:
        buyerName:
          type: string
        buyerPhone:
          type: string
        paymentMethod:
          type: string
          enum: [momo, card]
          default: momo
    Dispute:
      type: object
      required: [id, dealId, reason, status, createdAt]
      properties:
        id:
          type: integer
        dealId:
          type: integer
        reason:
          type: string
          enum: [item_never_arrived, wrong_damaged_item, incomplete_service]
        description:
          type: ["string", "null"]
        evidenceUrl:
          type: ["string", "null"]
        counterProofUrl:
          type: ["string", "null"]
        counterProofDescription:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, counter_submitted, resolved_refund, resolved_seller, escalated]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
...
    DisputeResolution:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [favor_buyer, favor_seller]
        resolution:
          type: string
...