# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from './dealStatus';
import type { Dispute } from './dispute';

export interface Deal {
  id: number;
  code: string;
  shortUrl: string;
  itemName: string;
  /** @nullable */
  description: string | null;
  price: number;
  feeAmount?: number;
  sellerPayout?: number;
  deliveryWindowHours: number;
  status: DealStatus;
  sellerId: number;
  /** @nullable */
  buyerPhone?: string | null;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  dispatchedAt?: string | null;
  /** @nullable */
  deliveryDeadline?: string | null;
  /** @nullable */
  settledAt?: string | null;
  dispute?: Dispute;
  createdAt: string;
}