# lib/api-zod/src/generated/types/dealInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DealInput {
  itemName: string;
  description?: string;
  price: number;
  deliveryWindowHours?: number;
}