# lib/api-zod/src/generated/types/dealList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Deal } from './deal';

export interface DealList {
  deals: Deal[];
  total: number;
}