# lib/api-zod/src/generated/types/dealPublic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DealPublicStatus } from './dealPublicStatus';

export interface DealPublic {
  id: number;
  code: string;
  itemName: string;
  /** @nullable */
  description?: string | null;
  price: number;
  deliveryWindowHours: number;
  status: DealPublicStatus;
  sellerName: string;
  createdAt: string;
}