# lib/api-zod/src/generated/types/dispute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DisputeReason } from './disputeReason';
import type { DisputeStatus } from './disputeStatus';

export interface Dispute {
  id: number;
  dealId: number;
  reason: DisputeReason;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  /** @nullable */
  counterProofUrl?: string | null;
  /** @nullable */
  counterProofDescription?: string | null;
  status: DisputeStatus;
  /** @nullable */
  resolution?: string | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}